requirements.txt
pandas
numpy
matplotlib
seaborn
python-chess
requests
tqdm
ipykernel
scikit-learn
